// src/app/models/user.ts
export class User {
    id: any;
    email: any;
    password: any;
    firstName: any;
    middleName: any;
    lastName: any;
    userTitle: any;
    photo: any;
    location_id: any
    location_name: any
    location_level_id: any
    passwordExpired: any
    requestPasswordChange: any
    role: any;
    role_id: any;
    roleTitle: any;
    permissions: any
    token: any;
    refreshToken: any;
    expireIn: any;
    gender: any;
    data: any;
}